refactor: refactor getters in HistoryAllPrintStatus
