+++ src/components/charts/HistoryAllPrintStatusTable.vue
@@ -14,26 +14,19 @@ import Component from 'vue-class-component'
 import { Mixins } from 'vue-property-decorator'
 import BaseMixin from '@/components/mixins/base'
 import { ServerHistoryStateAllPrintStatusEntry } from '@/store/server/history/types'
+import HistoryStatsMixin from '@/components/mixins/historyStats'
 
 @Component({
     components: {},
 })
-export default class HistoryAllPrintStatusTable extends Mixins(BaseMixin) {
+export default class HistoryAllPrintStatusTable extends Mixins(BaseMixin, HistoryStatsMixin) {
     get selectedJobs() {
         return this.$store.getters['server/history/getSelectedJobs']
     }
 
-    get allPrintStatusArray() {
-        return this.$store.getters['server/history/getAllPrintStatusArrayAll']
-    }
-
-    get selectedPrintStatusArray() {
-        return this.$store.getters['server/history/getSelectedPrintStatusArray']
-    }
-
     get printStatusArray() {
         const output: ServerHistoryStateAllPrintStatusEntry[] = []
-        const orgArray = this.selectedJobs.length ? this.selectedPrintStatusArray : this.allPrintStatusArray
+        const orgArray = this.selectedJobs.length ? this.selectedPrintStatusChartData : this.allPrintStatusChartData
 
         orgArray.forEach((status: ServerHistoryStateAllPrintStatusEntry) => {
             const tmp = { ...status }
