+++ src/store/server/history/getters.ts
@@ -77,132 +77,6 @@ export const getters: GetterTree<ServerHistoryState, any> = {
             : 0
     },
 
-    getAllPrintStatusArrayAll(state, getters, rootState) {
-        const output: ServerHistoryStateAllPrintStatusEntry[] = []
-
-        state.jobs.forEach((current) => {
-            const index = output.findIndex((element) => element.name === current.status)
-            if (index !== -1) output[index].value += 1
-            else {
-                const displayName = i18n.te(`History.StatusValues.${current.status}`, 'en')
-                    ? i18n.t(`History.StatusValues.${current.status}`).toString()
-                    : current.status
-
-                const itemStyle = {
-                    opacity: 0.9,
-                    color: '#424242',
-                    borderColor: '#1E1E1E',
-                    borderWidth: 2,
-                    borderRadius: 3,
-                }
-
-                switch (current.status) {
-                    case 'completed':
-                        itemStyle['color'] = '#BDBDBD'
-                        break
-
-                    case 'in_progress':
-                        itemStyle['color'] = '#EEEEEE'
-                        break
-
-                    case 'cancelled':
-                        itemStyle['color'] = '#616161'
-                        break
-                }
-
-                output.push({
-                    name: current.status,
-                    displayName,
-                    value: 1,
-                    itemStyle,
-                    showInTable: !rootState.gui?.view.history.hidePrintStatus.includes(current.status),
-                })
-            }
-        })
-
-        return output
-    },
-
-    getAllPrintStatusArray(state, getters) {
-        const output: ServerHistoryStateAllPrintStatusEntry[] = [...getters.getAllPrintStatusArrayAll]
-        const totalCount = state.jobs.length
-
-        const otherLimit = totalCount * 0.05
-        const others = output.filter((entry) => entry.value < otherLimit)
-        if (others.length > 1) {
-            let otherValue = 0
-
-            others.forEach((otherGroup) => {
-                const index = output.findIndex((entry) => entry.name === otherGroup.name)
-                if (index !== -1) {
-                    otherValue += output[index].value
-                    output.splice(index, 1)
-                }
-            })
-
-            output.push({
-                name: 'others',
-                displayName: i18n.t(`History.StatusValues.Others`).toString(),
-                value: otherValue,
-                itemStyle: {
-                    opacity: 0.9,
-                    color: '#616161',
-                    borderColor: '#1E1E1E',
-                    borderWidth: 2,
-                    borderRadius: 3,
-                },
-                showInTable: true,
-            })
-        }
-
-        return output
-    },
-
-    getSelectedPrintStatusArray(state, getters, rootState) {
-        const output: ServerHistoryStateAllPrintStatusEntry[] = []
-
-        getters.getSelectedJobs.forEach((current: ServerHistoryStateJob) => {
-            const index = output.findIndex((element) => element.name === current.status)
-            if (index !== -1) output[index].value += 1
-            else {
-                const displayName = i18n.te(`History.StatusValues.${current.status}`, 'en').toString()
-                    ? i18n.t(`History.StatusValues.${current.status}`).toString()
-                    : current.status
-                const itemStyle = {
-                    opacity: 0.9,
-                    color: '#424242',
-                    borderColor: '#1E1E1E',
-                    borderWidth: 2,
-                    borderRadius: 3,
-                }
-
-                switch (current.status) {
-                    case 'completed':
-                        itemStyle['color'] = '#BDBDBD'
-                        break
-
-                    case 'in_progress':
-                        itemStyle['color'] = '#EEEEEE'
-                        break
-
-                    case 'cancelled':
-                        itemStyle['color'] = '#616161'
-                        break
-                }
-
-                output.push({
-                    name: current.status,
-                    displayName,
-                    value: 1,
-                    itemStyle: itemStyle,
-                    showInTable: !rootState.gui?.view.history.hidePrintStatus.includes(current.status),
-                })
-            }
-        })
-
-        return output
-    },
-
     getFilamentUsageArray(state, getters) {
         // eslint-disable-next-line
         const output: any = []
